+++ src/misc/images.rs
@@ -1,7 +1,7 @@
 extern crate image;
 extern crate sixtyfps;
 
-use std::cmp::min;
+use std::cmp::max;
 
 use image::{imageops, DynamicImage, GenericImageView};
 
@@ -76,21 +76,10 @@ fn process_dynamic_image(
     max_width: u32,
     max_height: u32,
 ) -> ImageBuffer {
-    let width = cat_image.width();
-    let height = cat_image.height();
-    let ratio = width as f32 / height as f32;
-
-    let mut new_width = 0;
-    let mut new_height = 0;
-    if width > height && max_width > 0 {
-        new_width = min(width, max_width);
-        new_height = (new_width as f32 / ratio) as u32;
-    }
-    if height > width && max_height > 0 {
-        new_height = min(height, max_height);
-        new_width = (new_height as f32 * ratio) as u32;
-    }
-
+    let (new_width, new_height) = get_size(
+        (cat_image.width(), cat_image.height()),
+        (max_width, max_height),
+    );
     let cat_image = cat_image.resize(new_width, new_height, imageops::FilterType::Nearest);
 
     let cat_image = cat_image.into_rgba8();
@@ -100,4 +89,16 @@ fn process_dynamic_image(
         270 => image::imageops::rotate270(&cat_image),
         _ => cat_image,
     }
+}
+
+/// Get the actual size from the current size and the max size
+pub fn get_size((width, height): (u32, u32), (nwidth, nheight): (u32, u32)) -> (u32, u32) {
+    let wratio = nwidth as f32 / width as f32;
+    let hratio = nheight as f32 / height as f32;
+
+    let ratio = f32::min(wratio, hratio);
+
+    let nw = max((width as f32 * ratio).round() as u32, 1);
+    let nh = max((height as f32 * ratio).round() as u32, 1);
+    (nw, nh)
 }